Upgrade to Laravel 5.8

+++ database/migrations/2018_08_17_000000_create_products_table.php
@@ -14,11 +14,11 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('slug')->unique();
             $table->text('description');
-            $table->integer('price');
+            $table->bigInteger('price');
             $table->timestamps();
         });
     }
