+++ database/migrations/2018_08_17_000001_create_category_product_table.php
@@ -14,8 +14,8 @@ class CreateCategoryProductTable extends Migration
     public function up()
     {
         Schema::create('category_product', function (Blueprint $table) {
-            $table->integer('category_id')->unsigned()->index();
-            $table->integer('product_id')->unsigned()->index();
+            $table->bigInteger('category_id')->unsigned()->index();
+            $table->bigInteger('product_id')->unsigned()->index();
 
             $table->foreign('category_id')->references('id')->on('categories');
             $table->foreign('product_id')->references('id')->on('products');
