+++ database/migrations/2018_12_31_000002_create_shipping_methods_table.php
@@ -14,9 +14,9 @@ class CreateShippingMethodsTable extends Migration
     public function up()
     {
         Schema::create('shipping_methods', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
-            $table->integer('price');
+            $table->bigInteger('price');
             $table->timestamps();
         });
     }
