+++ database/migrations/2019_01_01_000000_create_payment_methods_table.php
@@ -14,8 +14,8 @@ class CreatePaymentMethodsTable extends Migration
     public function up()
     {
         Schema::create('payment_methods', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->string('card_type')->nullable();
             $table->string('last_four')->nullable();
             $table->boolean('default')->default(true);
